improve input component
