+++ src/pages/auth/login/index.tsx
@@ -1,6 +1,7 @@
 import { Input } from '@/components/Input';
 import { useAuth } from '@/hooks/useAuth';
 import { useTitle } from '@/hooks/useTitle';
+import { RegExp } from '@/plugins/constants/regExp';
 import { useAppSelector } from '@/store';
 import { FormEvent, useState } from 'react';
 import styles from './styles.module.scss';
@@ -10,9 +11,14 @@ export function LoginPage() {
   const auth = useAppSelector((state) => state.auth);
   const { login } = useAuth();
 
-  const [form, setForm] = useState({
+  const [form, setForm] = useState<{
+    email: string;
+    password: string;
+    errors: string[];
+  }>({
     email: '',
-    password: ''
+    password: '',
+    errors: []
   });
 
   const handleSubmit = async (e: FormEvent<HTMLFormElement>) => {
@@ -24,35 +30,39 @@ export function LoginPage() {
     <div>
       <p className={styles.something}>Login page</p>
       <form onSubmit={handleSubmit}>
-        <input
-          type='text'
-          placeholder='Email'
-          value={form.email}
-          onChange={(e) => setForm((prev) => ({ ...prev, email: e.target.value }))}
-        />
-        <input
-          type='password'
-          placeholder='Password'
-          value={form.password}
-          onChange={(e) => setForm((prev) => ({ ...prev, password: e.target.value }))}
-        />
-
         <Input
           type='email'
           label='Email'
           value={form.email}
-          onChange={(value) => setForm((prev) => ({ ...prev, email: value }))}
-          clearable
+          onChange={(value, valid) => {
+            setForm((prev) => ({
+              ...prev,
+              email: value,
+              errors: !valid ? [...prev.errors, 'email'] : prev.errors.filter((e) => e !== 'email')
+            }));
+          }}
+          regExp={RegExp.email}
+          errorMessage='Please enter a valid email address'
+          required
         />
         <Input
           type='password'
           value={form.password}
           label='Password'
-          onChange={(value) => setForm((prev) => ({ ...prev, password: value }))}
+          onChange={(value, valid) => {
+            setForm((prev) => ({
+              ...prev,
+              password: value,
+              errors: !valid ? [...prev.errors, 'email'] : prev.errors.filter((e) => e !== 'email')
+            }));
+          }}
+          required
+          regExp={RegExp.password}
+          errorMessage='Password must be at least 8 characters long and contain at least one number, one uppercase letter and one lowercase letter'
         />
 
         <button type='reset'>Clear</button>
-        <button type='submit' disabled={auth.loading}>
+        <button type='submit' disabled={auth.loading || form.errors.length > 0}>
           Login
         </button>
       </form>
