+++ src/plugins/constants/regExp.ts
@@ -1,0 +1,4 @@
+export const RegExp = {
+  email: /^[^\s@]+@[^\s@]+\.[^\s@]+$/,
+  password: /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d).{8,}$/
+};
